Replaced vector and unordered map with small vector and flat map. Added more optimizations to UBO logic.

+++ VK9-Library/CDevice9.cpp
@@ -83,7 +83,7 @@ CDevice9::CDevice9(C9* Instance, UINT Adapter, D3DDEVTYPE DeviceType, HWND hFocu
 	BOOST_LOG_TRIVIAL(info) << "GPU driver version: " << mDeviceProperties.driverVersion;
 	BOOST_LOG_TRIVIAL(info) << "GPU API version: " << mDeviceProperties.apiVersion;
 	
-	for (int32_t i = 0; i < mDeviceMemoryProperties.memoryHeapCount; i++)
+	for (size_t i = 0; i < mDeviceMemoryProperties.memoryHeapCount; i++)
 	{
 		if ((mDeviceMemoryProperties.memoryHeaps[i].flags & VK_MEMORY_HEAP_DEVICE_LOCAL_BIT) == VK_MEMORY_HEAP_DEVICE_LOCAL_BIT)
 		{
@@ -126,7 +126,7 @@ CDevice9::CDevice9(C9* Instance, UINT Adapter, D3DDEVTYPE DeviceType, HWND hFocu
 	VkExtensionProperties* extension = new VkExtensionProperties[extensionCount];
 	vkEnumerateDeviceExtensionProperties(mPhysicalDevice, nullptr, &extensionCount, extension);
 
-	for (int32_t i = 0; i < extensionCount; i++)
+	for (size_t i = 0; i < extensionCount; i++)
 	{
 		BOOST_LOG_TRIVIAL(info) << "CDevice9::CDevice9 extension available: " << extension[i].extensionName;	
 	}
@@ -229,7 +229,7 @@ CDevice9::CDevice9(C9* Instance, UINT Adapter, D3DDEVTYPE DeviceType, HWND hFocu
 	Otherwise look for one for presentation and one for graphics.
 	The index of the queue us stored for later use.
 	*/
-	for (int32_t i = 0; i < mQueueCount; i++)
+	for (size_t i = 0; i < mQueueCount; i++)
 	{
 		VkBool32 doesSupportPresentation = false;
 		VkBool32 doesSupportGraphics = false;
@@ -1033,7 +1033,7 @@ CDevice9::CDevice9(C9* Instance, UINT Adapter, D3DDEVTYPE DeviceType, HWND hFocu
 	VK_PRESENT_MODE_IMMEDIATE_KHR - Do not wait for vertical blanking to update the image.
 	VK_PRESENT_MODE_FIFO_KHR - Wait for the next vertical blanking interval to update the image. If the interval is missed wait for the next one. New images will be queued for display.
 	*/
-	for (int32_t i = 0; i < mPresentationModeCount; i++)
+	for (size_t i = 0; i < mPresentationModeCount; i++)
 	{
 		if (mPresentationModes[i] == VK_PRESENT_MODE_MAILBOX_KHR)
 		{
@@ -1128,7 +1128,7 @@ CDevice9::CDevice9(C9* Instance, UINT Adapter, D3DDEVTYPE DeviceType, HWND hFocu
 		BOOST_LOG_TRIVIAL(info) << "CDevice9::CDevice9 vkGetSwapchainImagesKHR succeeded.";
 	}
 
-	for (int32_t i = 0; i < mSwapchainImageCount; i++)
+	for (size_t i = 0; i < mSwapchainImageCount; i++)
 	{
 		VkImageViewCreateInfo color_image_view = {};
 
@@ -1168,7 +1168,7 @@ CDevice9::CDevice9(C9* Instance, UINT Adapter, D3DDEVTYPE DeviceType, HWND hFocu
 	commandBufferInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
 	commandBufferInfo.commandBufferCount = 1;
 
-	for (int32_t i = 0; i < mSwapchainImageCount; i++)
+	for (size_t i = 0; i < mSwapchainImageCount; i++)
 	{
 		mResult = vkAllocateCommandBuffers(mDevice, &commandBufferInfo, &mSwapchainBuffers[i]);
 		if (mResult != VK_SUCCESS)
@@ -1358,7 +1358,7 @@ CDevice9::CDevice9(C9* Instance, UINT Adapter, D3DDEVTYPE DeviceType, HWND hFocu
 
 	mFramebuffers = new VkFramebuffer[mSwapchainImageCount];
 
-	for (int32_t i = 0; i < mSwapchainImageCount; i++)
+	for (size_t i = 0; i < mSwapchainImageCount; i++)
 	{
 		attachments[0] = mSwapchainViews[i];
 		mResult = vkCreateFramebuffer(mDevice, &framebufferCreateInfo, nullptr, &mFramebuffers[i]);
@@ -1431,7 +1431,7 @@ CDevice9::~CDevice9()
 
 	mGarbageManager.DestroyHandles();
 
-	for (int32_t i = 0; i < mSwapChains.size(); i++)
+	for (size_t i = 0; i < mSwapChains.size(); i++)
 	{
 		delete mSwapChains[i];
 	}
@@ -1440,7 +1440,7 @@ CDevice9::~CDevice9()
 
 	if (mFramebuffers!= nullptr)
 	{
-		for (int32_t i = 0; i < mSwapchainImageCount; i++)
+		for (size_t i = 0; i < mSwapchainImageCount; i++)
 		{
 			if (mFramebuffers[i] != VK_NULL_HANDLE)
 			{
@@ -1493,7 +1493,7 @@ CDevice9::~CDevice9()
 
 	if (mSwapchainViews != nullptr)
 	{
-		for (int32_t i = 0; i < mSwapchainImageCount; i++)
+		for (size_t i = 0; i < mSwapchainImageCount; i++)
 		{
 			if (mSwapchainViews[i] != VK_NULL_HANDLE)
 			{
